Move functions to more appropriate files
