+++ bq.go
@@ -16,6 +16,9 @@ import (
 	"google.golang.org/api/option"
 
 	"github.com/mtgban/go-mtgban/mtgban"
+	"github.com/mtgban/go-mtgban/mtgmatcher"
+	"github.com/mtgban/go-mtgban/mtgstocks"
+	"github.com/mtgban/go-mtgban/tcgplayer"
 )
 
 type dbElement struct {
@@ -450,4 +453,105 @@ func updateScraper(group, tableName string) error {
 	}
 
 	return nil
+}
+
+func updateStaticData() {
+	if Infos == nil {
+		Infos = map[string]mtgban.InventoryRecord{}
+	}
+
+	SealedEditionsSorted, SealedEditionsList = getSealedEditions()
+	AllEditionsKeys, AllEditionsMap = getAllEditions()
+	TreeEditionsKeys, TreeEditionsMap = getTreeEditions()
+
+	TotalSets = len(AllEditionsKeys)
+	TotalUnique = len(mtgmatcher.GetUUIDs())
+	var totalCards int
+	for _, key := range AllEditionsKeys {
+		totalCards += AllEditionsMap[key].Size
+	}
+	TotalCards = totalCards
+
+	if !SkipInitialRefresh {
+		go loadInfos()
+	}
+	go runSealedAnalysis()
+
+	// Load prices for API users
+	if !DevMode {
+		go prepareCKAPI()
+	}
+}
+
+func loadInfos() {
+	log.Println("Loading infos")
+
+	for _, seller := range []mtgban.Seller{
+		mtgstocks.NewScraper(),
+		mtgstocks.NewScraperIndex(),
+		tcgplayer.NewScraperSYP(),
+	} {
+		loadInfoScraper(seller)
+	}
+	ServerNotify("refresh", "infos refreshed")
+}
+
+func loadInfoScraper(seller mtgban.Seller) {
+	inv, err := seller.Inventory()
+	if err != nil {
+		log.Println(err)
+		return
+	}
+	Infos[seller.Info().Shorthand] = inv
+	log.Println("Infos loaded:", seller.Info().Name)
+}
+
+func loadDatastore() error {
+	allPrintingsReader, err := os.Open(AllPrintingsFileName)
+	if err != nil {
+		return err
+	}
+	defer allPrintingsReader.Close()
+
+	return mtgmatcher.LoadDatastore(allPrintingsReader)
+}
+
+func loadSellerFromFile(fname string) (mtgban.Seller, error) {
+	file, err := os.Open(fname)
+	if err != nil {
+		return nil, err
+	}
+	defer file.Close()
+
+	return mtgban.ReadSellerFromJSON(file)
+}
+
+func dumpSellerToFile(seller mtgban.Seller, fname string) error {
+	file, err := os.Create(fname)
+	if err != nil {
+		return err
+	}
+	defer file.Close()
+
+	return mtgban.WriteSellerToJSON(seller, file)
+}
+
+func loadVendorFromFile(fname string) (mtgban.Vendor, error) {
+	file, err := os.Open(fname)
+	if err != nil {
+		return nil, err
+	}
+	defer file.Close()
+
+	return mtgban.ReadVendorFromJSON(file)
+}
+
+func dumpVendorToFile(vendor mtgban.Vendor, fname string) error {
+	file, err := os.Create(fname)
+	if err != nil {
+		return err
+	}
+	defer file.Close()
+
+	return mtgban.WriteVendorToJSON(vendor, file)
 }
